Fixed GNU warning

+++ src/ptclib/pstun.cxx
@@ -484,7 +484,7 @@ PSTUNAttribute * PSTUNMessage::GetFirstAttribute() const
     return NULL;
 
   int length = ((PSTUNMessageHeader *)theArray)->msgLength;
-  if (length < sizeof(PSTUNAttribute))
+  if (length < (int)sizeof(PSTUNAttribute))
     return NULL;
 
   if (GetSize() < ((int)sizeof(PSTUNMessageHeader) + length))
@@ -530,7 +530,7 @@ bool PSTUNMessage::IsValid() const
 
   // check attributes
   PSTUNAttribute * attrib = GetFirstAttribute();
-  while (attrib != NULL && length >= sizeof(PSTUNAttribute)) {
+  while (attrib != NULL && length >= (int)sizeof(PSTUNAttribute)) {
     length -= CalcPaddedAttributeLength(attrib->length);
     attrib = attrib->GetNext();
   }
